+++ IceBuilder_VS2010/Properties/AssemblyInfo.cs
@@ -1,0 +1,19 @@
+// **********************************************************************
+//
+// Copyright (c) 2009-2016 ZeroC, Inc. All rights reserved.
+//
+// **********************************************************************
+
+using System;
+using System.Reflection;
+using System.Resources;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("IceBuilderVS2010")]
+[assembly: AssemblyDescription("Ice Builder Visual Studio Extension")]
+[assembly: AssemblyCompany("ZeroC, Inc.")]
+[assembly: AssemblyProduct("Ice Builder")]
+[assembly: AssemblyCopyright("Copyright (c) 2009-2017 ZeroC, Inc.")]
+[assembly: AssemblyVersion("4.3.7.0")]
+[assembly: ComVisibleAttribute(false)]