+++ SignExtensionTask/Properties/AssemblyInfo.cs
@@ -1,0 +1,20 @@
+// **********************************************************************
+//
+// Copyright (c) 2009-2015 ZeroC, Inc. All rights reserved.
+//
+// **********************************************************************
+
+using System;
+using System.Reflection;
+using System.Resources;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("IceBuilder.SignTasks")]
+[assembly: AssemblyDescription("Ice Builder Sign Tasks")]
+[assembly: AssemblyCompany("ZeroC, Inc.")]
+[assembly: AssemblyProduct("Ice Builder")]
+[assembly: AssemblyCopyright("Copyright (c) 2009-2015 ZeroC, Inc.")]
+[assembly: AssemblyVersion("4.0.0")]
+[assembly: AssemblyDelaySign(false)]
+[assembly: ComVisibleAttribute(false)]